// src/@types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IArticleFields {
  /** Title */
  title?: string | undefined;

  /** Slug */
  slug?: string | undefined;

  /** Description */
  description?: string | undefined;

  /** Publish Date */
  publishDate?: string | undefined;

  /** Content */
  content: Document;
}

export interface IArticle extends Entry<IArticleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'article';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IAuthorFields {
  /** Name */
  name: string;

  /** Image */
  image?: Asset | undefined;

  /** Date */
  date?: string | undefined;
}

/** Author content type */

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'author';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IBlogPostFields {
  /** Title */
  title?: string | undefined;

  /** Publish Date */
  publishDate?: string | undefined;

  /** Slug */
  slug?: string | undefined;

  /** Feature Image */
  featureImage?: Asset | undefined;

  /** Author */
  author?: IAuthor | undefined;

  /** Category */
  category?: ICategory | undefined;

  /**  Excerpt */
  excerpt?: string | undefined;

  /** Content */
  content?: Document | undefined;
}

/** Blog post content type */

export interface IBlogPost extends Entry<IBlogPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'blogPost';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICategoryFields {
  /** Category Name */
  categoryName?: string | undefined;

  /** Description */
  description?: string | undefined;
}

/** Category content type */

export interface ICategory extends Entry<ICategoryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'category';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'article' | 'author' | 'blogPost' | 'category';

export type IEntry = IArticle | IAuthor | IBlogPost | ICategory;

export type LOCALE_CODE = 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';